requirements.txt
pillow
pytesseract
mysql-connector-python
opencv-python
numpy
pyside6
